requirements.txt
lxml
python-twitter